Created taskUpdateMenu.js
